// src/frontend/src/backend.d.ts
import type { Principal } from "@icp-sdk/core/principal";
export interface Some<T> {
    __kind__: "Some";
    value: T;
}
export interface None {
    __kind__: "None";
}
export type Option<T> = Some<T> | None;
export interface Settings {
    vibrationEnabled: boolean;
    ndModeEnabled: boolean;
    audioVolume: bigint;
    vibrationStrength: bigint;
    audioEnabled: boolean;
}
export interface backendInterface {
    getSettings(): Promise<Settings>;
    updateSettings(newSettings: Settings): Promise<void>;
}